Add table group & minor changes

+++ DB/test_read_cfg.sql
@@ -222,7 +222,7 @@ BEGIN
         IF COALESCE(value_var, '') = '' THEN 
             value_var = '0' || value_var;
         END IF;
-        UPDATE comtrade.cfg SET samp = array_append(samp::int[], value_var::int) WHERE name=name_var;
+        UPDATE comtrade.cfg SET samp = array_append(samp::float[], value_var::float) WHERE name=name_var;
 
         value_var = trim(values_var[2]); 
         IF COALESCE(value_var, '') = '' THEN 
@@ -285,13 +285,13 @@ BEGIN
         IF COALESCE(value_var, '') = '' THEN
             value_var = trim(value_var);
         END IF;
-        UPDATE comtrade.cfg SET time_code = value_var::float WHERE name=name_var;
+        UPDATE comtrade.cfg SET time_code = value_var WHERE name=name_var;
 
         value_var = trim(values_var[2]);
         IF COALESCE(value_var, '') = '' THEN
             value_var = trim(value_var);
         END IF;
-        UPDATE comtrade.cfg SET local_code = value_var::float WHERE name=name_var;
+        UPDATE comtrade.cfg SET local_code = value_var WHERE name=name_var;
 
         file_content_var = right(file_content_var, char_length(file_content_var) - position(Chr(13) || Chr(10) in file_content_var) - 1);
 
@@ -302,13 +302,13 @@ BEGIN
         IF COALESCE(value_var, '') = '' THEN
             value_var = trim(value_var);
         END IF;
-        UPDATE comtrade.cfg SET tmq_code = value_var::float WHERE name=name_var;
+        UPDATE comtrade.cfg SET tmq_code = value_var WHERE name=name_var;
 
         value_var = trim(values_var[2]);
         IF COALESCE(value_var, '') = '' THEN
             value_var = trim(value_var);
         END IF;
-        UPDATE comtrade.cfg SET leapsec = value_var::float WHERE name=name_var;
+        UPDATE comtrade.cfg SET leapsec = value_var::int WHERE name=name_var;
     END IF;
 
     -- RAISE EXCEPTION 'line_var="%"; values_var="%"; value_var="%"; int_a_var="%", int_d_var="%".', line_var, values_var, value_var, int_a_var, int_d_var;
